Add UInt128Converter

+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Value/ByteConverter.cs
@@ -17,7 +17,9 @@ namespace System.Text.Json.Serialization.Converters
 
         public override void Write(Utf8JsonWriter writer, byte value, JsonSerializerOptions options)
         {
-            writer.WriteNumberValue(value);
+            // TODO: [ActiveIssue("https://github.com/dotnet/runtime/issues/74001")]
+            // TODO: Casting to long should not be required
+            writer.WriteNumberValue((long)value);
         }
 
         internal override byte ReadAsPropertyNameCore(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
@@ -27,7 +29,9 @@ namespace System.Text.Json.Serialization.Converters
 
         internal override void WriteAsPropertyNameCore(Utf8JsonWriter writer, byte value, JsonSerializerOptions options, bool isWritingExtensionDataProperty)
         {
-            writer.WritePropertyName(value);
+            // TODO: [ActiveIssue("https://github.com/dotnet/runtime/issues/74001")]
+            // TODO: Casting to long should not be required
+            writer.WritePropertyName((long)value);
         }
 
         internal override byte ReadNumberWithCustomHandling(ref Utf8JsonReader reader, JsonNumberHandling handling, JsonSerializerOptions options)
@@ -42,13 +46,15 @@ namespace System.Text.Json.Serialization.Converters
 
         internal override void WriteNumberWithCustomHandling(Utf8JsonWriter writer, byte value, JsonNumberHandling handling)
         {
+            // TODO: [ActiveIssue("https://github.com/dotnet/runtime/issues/74001")]
+            // TODO: Casting to long should not be required
             if ((JsonNumberHandling.WriteAsString & handling) != 0)
             {
-                writer.WriteNumberValueAsString(value);
+                writer.WriteNumberValueAsString((long)value);
             }
             else
             {
-                writer.WriteNumberValue(value);
+                writer.WriteNumberValue((long)value);
             }
         }
     }
