+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Value/UInt16Converter.cs
@@ -28,7 +28,9 @@ namespace System.Text.Json.Serialization.Converters
 
         internal override void WriteAsPropertyNameCore(Utf8JsonWriter writer, ushort value, JsonSerializerOptions options, bool isWritingExtensionDataProperty)
         {
-            writer.WritePropertyName(value);
+            // TODO: [ActiveIssue("https://github.com/dotnet/runtime/issues/74001")]
+            // TODO: Casting to long should not be required
+            writer.WritePropertyName((long)value);
         }
 
         internal override ushort ReadNumberWithCustomHandling(ref Utf8JsonReader reader, JsonNumberHandling handling, JsonSerializerOptions options)
@@ -46,7 +48,9 @@ namespace System.Text.Json.Serialization.Converters
         {
             if ((JsonNumberHandling.WriteAsString & handling) != 0)
             {
-                writer.WriteNumberValueAsString(value);
+                // TODO: [ActiveIssue("https://github.com/dotnet/runtime/issues/74001")]
+                // TODO: Casting to long should not be required
+                writer.WriteNumberValueAsString((long)value);
             }
             else
             {
